+++ core-ng/src/main/java/core/framework/impl/kafka/MessageProducer.java
@@ -6,5 +6,5 @@ import org.apache.kafka.clients.producer.ProducerRecord;
  * @author neo
  */
 public interface MessageProducer {
-    void send(ProducerRecord<String, byte[]> record);
+    void send(ProducerRecord<byte[], byte[]> record);
 }