+++ core-ng/src/main/java/core/framework/impl/kafka/MessageProducerImpl.java
@@ -6,7 +6,6 @@ import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.serialization.ByteArraySerializer;
-import org.apache.kafka.common.serialization.StringSerializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,7 +21,7 @@ public class MessageProducerImpl implements MessageProducer {
     private final Logger logger = LoggerFactory.getLogger(MessageProducerImpl.class);
     private final String uri;
     private final String name;
-    private final Producer<String, byte[]> producer;
+    private final Producer<byte[], byte[]> producer;
 
     public MessageProducerImpl(String uri, String name) {
         var watch = new StopWatch();
@@ -33,7 +32,8 @@ public class MessageProducerImpl implements MessageProducer {
             Map<String, Object> config = Map.of(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, uri,  // immutable map requires value must not be null
                     ProducerConfig.COMPRESSION_TYPE_CONFIG, "snappy",
                     ProducerConfig.MAX_BLOCK_MS_CONFIG, Duration.ofSeconds(30).toMillis());  // metadata update timeout
-            this.producer = new KafkaProducer<>(config, new StringSerializer(), new ByteArraySerializer());
+            var serializer = new ByteArraySerializer();
+            this.producer = new KafkaProducer<>(config, serializer, serializer);
             producerMetrics.set(producer.metrics());
         } finally {
             logger.info("create kafka producer, uri={}, name={}, elapsed={}", uri, name, watch.elapsed());
@@ -41,7 +41,7 @@ public class MessageProducerImpl implements MessageProducer {
     }
 
     @Override
-    public void send(ProducerRecord<String, byte[]> record) {
+    public void send(ProducerRecord<byte[], byte[]> record) {
         producer.send(record);
     }