+++ core-ng/src/main/java/core/framework/impl/log/KafkaAppender.java
@@ -6,6 +6,8 @@ import core.framework.impl.log.filter.LogFilter;
 import core.framework.impl.log.message.ActionLogMessage;
 import core.framework.impl.log.message.LogTopics;
 import core.framework.impl.log.message.StatMessage;
+import core.framework.util.StopWatch;
+import core.framework.util.Strings;
 import core.framework.util.Threads;
 import org.apache.kafka.clients.producer.Callback;
 import org.apache.kafka.clients.producer.KafkaProducer;
@@ -13,7 +15,6 @@ import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.serialization.ByteArraySerializer;
-import org.apache.kafka.common.serialization.StringSerializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -29,15 +30,15 @@ import java.util.concurrent.atomic.AtomicBoolean;
  */
 public final class KafkaAppender implements Appender {
     public final ProducerMetrics producerMetrics = new ProducerMetrics("log-forwarder");
-    private final BlockingQueue<Object> queue = new LinkedBlockingQueue<>();
     private final Logger logger = LoggerFactory.getLogger(KafkaAppender.class);
+    private final BlockingQueue<Object> queue = new LinkedBlockingQueue<>();
     private final String appName;
-    private final Producer<String, byte[]> producer;
+    private final Producer<byte[], byte[]> producer;
 
     private final AtomicBoolean stop = new AtomicBoolean(false);
     private final Thread logForwarderThread;
-    private final JSONWriter<ActionLogMessage> actionLogWriter = JSONWriter.of(ActionLogMessage.class);
-    private final JSONWriter<StatMessage> statWriter = JSONWriter.of(StatMessage.class);
+    private final JSONWriter<ActionLogMessage> actionLogWriter;
+    private final JSONWriter<StatMessage> statWriter;
     private final Callback callback = (metadata, exception) -> {
         if (exception != null) {
             logger.warn("failed to send log message", exception);
@@ -46,16 +47,23 @@ public final class KafkaAppender implements Appender {
     };
 
     public KafkaAppender(String uri, String appName) {
+        var watch = new StopWatch();
         this.appName = appName;
-
-        Map<String, Object> config = Map.of(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, uri,
-                ProducerConfig.ACKS_CONFIG, "0",                                    // no acknowledge to maximize performance
-                ProducerConfig.MAX_BLOCK_MS_CONFIG, Duration.ofSeconds(30).toMillis(),  // metadata update timeout
-                ProducerConfig.COMPRESSION_TYPE_CONFIG, "snappy",
-                ProducerConfig.LINGER_MS_CONFIG, 50,
-                ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder");
-        producer = new KafkaProducer<>(config, new StringSerializer(), new ByteArraySerializer());
-        producerMetrics.set(producer.metrics());
+        actionLogWriter = JSONWriter.of(ActionLogMessage.class);
+        statWriter = JSONWriter.of(StatMessage.class);
+        try {
+            Map<String, Object> config = Map.of(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, uri,
+                    ProducerConfig.ACKS_CONFIG, "0",                                    // no acknowledge to maximize performance
+                    ProducerConfig.MAX_BLOCK_MS_CONFIG, Duration.ofSeconds(30).toMillis(),  // metadata update timeout
+                    ProducerConfig.COMPRESSION_TYPE_CONFIG, "snappy",
+                    ProducerConfig.LINGER_MS_CONFIG, 50,
+                    ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder");      // if not specify, kafka uses producer-${seq} name, also impact jmx naming
+            var serializer = new ByteArraySerializer();
+            producer = new KafkaProducer<>(config, serializer, serializer);
+            producerMetrics.set(producer.metrics());
+        } finally {
+            logger.info("create log forwarder, uri={}, elapsed={}", uri, watch.elapsed());
+        }
 
         logForwarderThread = new Thread(() -> {
             logger.info("log forwarder thread started, uri={}", uri);
@@ -63,9 +71,11 @@ public final class KafkaAppender implements Appender {
                 try {
                     Object message = queue.take();
                     if (message instanceof ActionLogMessage) {
-                        producer.send(new ProducerRecord<>(LogTopics.TOPIC_ACTION_LOG, actionLogWriter.toJSON((ActionLogMessage) message)), callback);
+                        var actionLog = (ActionLogMessage) message;
+                        producer.send(new ProducerRecord<>(LogTopics.TOPIC_ACTION_LOG, Strings.bytes(actionLog.id), actionLogWriter.toJSON(actionLog)), callback);
                     } else if (message instanceof StatMessage) {
-                        producer.send(new ProducerRecord<>(LogTopics.TOPIC_STAT, statWriter.toJSON((StatMessage) message)), callback);
+                        var stat = (StatMessage) message;
+                        producer.send(new ProducerRecord<>(LogTopics.TOPIC_STAT, Strings.bytes(stat.id), statWriter.toJSON(stat)), callback);
                     }
                 } catch (Throwable e) {
                     if (!stop.get()) {